+++ .eslintrc.js
@@ -1,0 +1,5 @@
+require('ts-node').register()
+
+module.exports = {
+  extends: '@1stg',
+}